+++ src/x86_64.rs
@@ -6,7 +6,8 @@
 use arrayvec::ArrayVec;
 use std::ops::ControlFlow;
 use thiserror::Error;
-use zerocopy::{FromBytes, FromZeroes, Ref, LE};
+use zerocopy::{FromBytes, Ref, Unaligned, LE};
+use zerocopy_derive::{FromBytes, FromZeroes, Unaligned};
 
 type U16 = zerocopy::U16<LE>;
 
@@ -20,7 +21,7 @@ pub struct FunctionTableEntries<'a> {
 }
 
 /// A runtime function record in the function table.
-#[derive(FromBytes, FromZeroes, Debug, Clone, Copy)]
+#[derive(Unaligned, FromZeroes, FromBytes, Debug, Clone, Copy)]
 #[repr(C)]
 pub struct RuntimeFunction {
     /// The start relative virtual address of the function.
@@ -45,7 +46,7 @@ impl<'a> FunctionTableEntries<'a> {
     /// Get the `RuntimeFunction`s in the function table, if the parsed data is well-aligned and
     /// sized.
     pub fn functions(&self) -> Option<&'a [RuntimeFunction]> {
-        Ref::new_slice(self.data).map(|lv| lv.into_slice())
+        Ref::new_slice_unaligned(self.data).map(|lv| lv.into_slice())
     }
 
     /// Lookup the runtime function that contains the given relative virtual address.
@@ -159,7 +160,7 @@ impl<'a> Iterator for FunctionTableEntries<'a> {
     type Item = &'a RuntimeFunction;
 
     fn next(&mut self) -> Option<Self::Item> {
-        let (rf, rest) = Ref::<_, RuntimeFunction>::new_from_prefix(self.data)?;
+        let (rf, rest) = Ref::<_, RuntimeFunction>::new_unaligned_from_prefix(self.data)?;
         self.data = rest;
         Some(rf.into_ref())
     }
@@ -173,12 +174,13 @@ struct Sections<'a> {
 
 impl<'a> Sections<'a> {
     pub fn parse(image: &'a [u8]) -> Option<Self> {
-        let sig_offset = Ref::<_, U32>::new(image.get(0x3c..0x40)?)?.get() as usize;
+        let sig_offset = Ref::<_, U32>::new_unaligned(image.get(0x3c..0x40)?)?.get() as usize;
         // Offset to the COFF header
         let coff_image = image.get(sig_offset + 4..)?;
-        let section_count = Ref::<_, U16>::new(coff_image.get(2..4)?)?.get() as usize;
-        let size_of_optional_header = Ref::<_, U16>::new(coff_image.get(16..18)?)?.get() as usize;
-        let sections = Ref::<_, [Section]>::new_slice_from_prefix(
+        let section_count = Ref::<_, U16>::new_unaligned(coff_image.get(2..4)?)?.get() as usize;
+        let size_of_optional_header =
+            Ref::<_, U16>::new_unaligned(coff_image.get(16..18)?)?.get() as usize;
+        let sections = Ref::<_, [Section]>::new_slice_unaligned_from_prefix(
             &coff_image[20 + size_of_optional_header..],
             section_count,
         )?
@@ -210,7 +212,8 @@ impl<'a> Sections<'a> {
     }
 }
 
-#[derive(FromBytes, FromZeroes, Debug, Clone, Copy)]
+#[derive(Unaligned, FromZeroes, FromBytes, Debug, Clone, Copy)]
+#[repr(C)]
 struct Section {
     _name: [u8; 8],
     virtual_size: U32,
@@ -328,7 +331,7 @@ impl TryFrom<u8> for XmmRegister {
 /// Fixed data at the start of [PE UnwindInfo][unwindinfo].
 ///
 /// [unwindinfo]: https://learn.microsoft.com/en-us/cpp/build/exception-handling-x64?view=msvc-170#struct-unwind_info
-#[derive(FromBytes, FromZeroes, Debug, Clone, Copy)]
+#[derive(Unaligned, FromZeroes, FromBytes, Debug, Clone, Copy)]
 #[repr(C)]
 pub struct UnwindInfoHeader {
     /// The unwind information version and flags.
@@ -704,12 +707,12 @@ impl<'a> UnwindInfo<'a> {
     ///
     /// Returns None if there aren't enough bytes or the alignment is incorrect.
     pub fn parse(data: &'a [u8]) -> Option<Self> {
-        let (header, rest) = Ref::<_, UnwindInfoHeader>::new_from_prefix(data)?;
+        let (header, rest) = Ref::<_, UnwindInfoHeader>::new_unaligned_from_prefix(data)?;
         if header.version() != 1 {
             return None;
         }
         let (unwind_codes, rest) =
-            Ref::new_slice_from_prefix(rest, header.unwind_codes_len as usize * 2)?;
+            Ref::new_slice_unaligned_from_prefix(rest, header.unwind_codes_len as usize * 2)?;
         Some(UnwindInfo {
             header: header.into_ref(),
             unwind_codes: unwind_codes.into_slice(),
@@ -726,20 +729,22 @@ impl<'a> UnwindInfo<'a> {
     pub fn trailer(&self) -> Option<UnwindInfoTrailer<'a>> {
         let flags = self.flags();
         if flags.contains(UnwindInfoFlags::EHANDLER) {
-            let (handler_address, handler_data) = Ref::<_, U32>::new_from_prefix(self.rest)?;
+            let (handler_address, handler_data) =
+                Ref::<_, U32>::new_unaligned_from_prefix(self.rest)?;
             Some(UnwindInfoTrailer::ExceptionHandler {
                 handler_address: handler_address.into_ref(),
                 handler_data,
             })
         } else if flags.contains(UnwindInfoFlags::UHANDLER) {
-            let (handler_address, handler_data) = Ref::<_, U32>::new_from_prefix(self.rest)?;
+            let (handler_address, handler_data) =
+                Ref::<_, U32>::new_unaligned_from_prefix(self.rest)?;
             Some(UnwindInfoTrailer::TerminationHandler {
                 handler_address: handler_address.into_ref(),
                 handler_data,
             })
         } else if flags.contains(UnwindInfoFlags::CHAININFO) {
             Some(UnwindInfoTrailer::ChainedUnwindInfo {
-                chained: Ref::<_, RuntimeFunction>::new(self.rest)?.into_ref(),
+                chained: Ref::<_, RuntimeFunction>::new_unaligned(self.rest)?.into_ref(),
             })
         } else {
             None
@@ -769,8 +774,8 @@ impl<'a> UnwindOperations<'a> {
         c.read::<UnwindCode>()
     }
 
-    fn read<T: FromBytes>(&mut self) -> Option<&'a T> {
-        let (v, rest) = Ref::<_, T>::new_from_prefix(self.0)?;
+    fn read<T: Unaligned + FromBytes>(&mut self) -> Option<&'a T> {
+        let (v, rest) = Ref::<_, T>::new_unaligned_from_prefix(self.0)?;
         self.0 = rest;
         Some(v.into_ref())
     }
@@ -861,7 +866,7 @@ pub enum UnwindOperationCode {
 }
 
 /// A single step to unwind operations done in a frame's prolog.
-#[derive(FromBytes, FromZeroes, Debug, Clone, Copy)]
+#[derive(Unaligned, FromZeroes, FromBytes, Debug, Clone, Copy)]
 #[repr(C)]
 pub struct UnwindCode {
     /// The byte offset into the prolog where the operation was done.